Add info about term effects to data set

Adds the contribution (plus confidence intervals) of a specific term to the linear predictor to the provided data. Largely a wrapper to predict.gam, with type="terms". Thus most arguments and their documentation below is from predict.gam.

add_term(newdata, object, term, se.fit = TRUE, type = "terms",
  se.mult = 2, relative = FALSE, ...)

Arguments

newdata
A data frame or list containing the values of the model covariates at which predictions are required. If this is not provided then predictions corresponding to the original data are returned. If newdata is provided then it should contain all the variables needed for prediction: a warning is generated if not. See details for use with link{linear.functional.terms}.

object
a fitted gam object as produced by gam().

term
A character (vector) or regular expression indicating for which term(s) information should be extracted and added to data set.

se.fit
when this is TRUE (not default) standard error estimates are returned for each prediction.

type
When this has the value "link" (default) the linear predictor (possibly with associated standard errors) is returned. When type="terms" each component of the linear predictor is returned seperately (possibly with standard errors): this includes parametric model components, followed by each smooth component, but excludes any offset and any intercept. type="iterms" is the same, except that any standard errors returned for smooth components will include the uncertainty about the intercept/overall mean. When type="response" predictions on the scale of the response are returned (possibly with approximate standard errors). When type="lpmatrix" then a matrix is returned which yields the values of the linear predictor (minus any offset) when postmultiplied by the parameter vector (in this case se.fit is ignored). The latter option is most useful for getting variance estimates for quantities derived from the model: for example integrated quantities, or derivatives of smooths. A linear predictor matrix can also be used to implement approximate prediction outside R (see example code, below).

se.mult
The factor by which standard errors are multiplied to form confidence intervals.

relative
If TRUE, calculates relative risk contribution, that is \((X-\bar{X})'\beta\) and respective confidence intervals if se.fit = TRUE. Defaults to FALSE.

...
Further arguments passed to predict.gam

See also

predict.gam, add_hazard

Examples

library(mgcv)
data("veteran", package="survival")
ped <- split_data(Surv(time, status)~ age, data=veteran, id="id")
pam <- gam(ped_status ~ s(tend),
        data = ped, family=poisson(), offset=offset)
pinf <- ped_info(ped)
add_term(pinf, pam, term="tend")

#> # A tibble: 97 x 9
#>    tstart  tend intlen intmid interval   age    fit    low  high
#>     <dbl> <dbl>  <dbl>  <dbl> <fctr>   <dbl>  <dbl>  <dbl> <dbl>
#>  1   0     1.00   1.00  0.500 (0,1]     58.2 0.0969 0.0244 0.169
#>  2   1.00  2.00   1.00  1.50  (1,2]     58.2 0.0953 0.0240 0.167
#>  3   2.00  3.00   1.00  2.50  (2,3]     58.2 0.0938 0.0237 0.164
#>  4   3.00  4.00   1.00  3.50  (3,4]     58.2 0.0922 0.0233 0.161
#>  5   4.00  7.00   3.00  5.50  (4,7]     58.2 0.0874 0.0221 0.153
#>  6   7.00  8.00   1.00  7.50  (7,8]     58.2 0.0858 0.0217 0.150
#>  7   8.00 10.0    2.00  9.00  (8,10]    58.2 0.0826 0.0209 0.144
#>  8  10.0  11.0    1.00 10.5   (10,11]   58.2 0.0811 0.0205 0.142
#>  9  11.0  12.0    1.00 11.5   (11,12]   58.2 0.0795 0.0201 0.139
#> 10  12.0  13.0    1.00 12.5   (12,13]   58.2 0.0779 0.0197 0.136
#> # ... with 87 more rows