creates one instance of Lag/Lead mat

creates one instance of Lag/Lead mat

create_Lmat(te, cut, t.lag = 1, lead.const = NULL, te.add.lag = 0,
  te.add.lead = 0, lead.factor = 0, t.min = 0)

Arguments

te
Numeric/Integer vector specifying the times at which exposure occurred.

cut
The break points dividing the follow up into intervals.

t.lag
A numeric constant, specifying the time (from te) before te can affect hazard.

lead.const
A numeric constant, specifying the constant amount of time after te + t.lag, in which te can still affect hazard.

te.add.lag
A numeric constant added to te before application of lag time

te.add.lead
A numeric constant, added to te before application of lead time.

lead.factor
If the lead time is dynamic, this factor can be set different to zero, such that t.lead=lead.const + lead.factor*te.

t.min
If some intervals are not of interest only intervals for t > t.min are returned.

Value

A data frame with intervals as first column and length(te) columns specifying the lag/lead for each te.