Extract interval information and median/modus values for covariates

Given an object of class ped, returns data frame with one row for each interval containing interval information, median values for numerical variables and modi for non-numeric variables in the data set.

ped_info(ped)

Arguments

ped
An object of class ped as returned by split_data.

Value

A data frame with one row for each interval in ped.

See also

int_info, sample_info

Examples

data("veteran", package="survival")
ped <- split_data(Surv(time, status)~ trt + age, data=veteran, id="id")
ped_info(ped) # note that trt is coded 1/2, should be fixed beforehand

#> # A tibble: 97 x 7
#>    tstart  tend intlen intmid interval   trt   age
#>     <dbl> <dbl>  <dbl>  <dbl> <fctr>   <dbl> <dbl>
#>  1   0     1.00   1.00  0.500 (0,1]     1.47  58.2
#>  2   1.00  2.00   1.00  1.50  (1,2]     1.47  58.2
#>  3   2.00  3.00   1.00  2.50  (2,3]     1.47  58.2
#>  4   3.00  4.00   1.00  3.50  (3,4]     1.47  58.2
#>  5   4.00  7.00   3.00  5.50  (4,7]     1.47  58.2
#>  6   7.00  8.00   1.00  7.50  (7,8]     1.47  58.2
#>  7   8.00 10.0    2.00  9.00  (8,10]    1.47  58.2
#>  8  10.0  11.0    1.00 10.5   (10,11]   1.47  58.2
#>  9  11.0  12.0    1.00 11.5   (11,12]   1.47  58.2
#> 10  12.0  13.0    1.00 12.5   (12,13]   1.47  58.2
#> # ... with 87 more rows